ciscenje koda i swagger
