+++ trading-service/controllers/orders/ExecuteOrders.go
@@ -1,6 +1,5 @@
 package orders
 
-/
 /*
 
 import (
