+++ trading-service/cron/CronJobs.go
@@ -1,7 +1,6 @@
 package cron
 
 import (
-	"banka1.com/controllers/orders"
 	"banka1.com/db"
 	"encoding/json"
 	"errors"
